Removed home_position function
The function to move to a 'home' position would tend to cause the robot to be suck in a singularity, making it unable to move to a second position when given a cartesian pose; requiring manual movement out of the home position. This function caused more harm than it was worth, so it will be removed.
Currently it is still callable, it will just give you a message saying it will be removed and do nothing else.